Wix storybook utils/two point oh (#559)
* chore(FormComponents): move option component to file

* chore(FormComponents): move under AutoExample

these are not to be exported but import-path does that
and someone might think these are public api. well they're not, these
are only for AutoExample

* chore(SectionCollapse): add collapsable component for props list

* chore(AutoExample): make use of SectionCollapse

* feat(categorizeProps): add function for object splitting

* chore(AutoExample): use categorizeProps to split props into sections

* chore(*): remove old Story * ComponentMetaInfoGetter

previously we had Story working with ComponentMetaInfoGetter which was
to be deprecated for a long time (it was working during run time)

the current approach is parsing component during build time with
react-autodocs-utils. it makes Story and ComponentMetaInfoGetter
redundant and requires StoryNew to be renamed to just Story

* chore(AutoExample): cleanup runtime parsing

* chore(categorizeProps): don't add unnecesarry properties

* chore(package): use react-autodocs-utils 3.0.0

* chore(package): bump version to 2.0.0

* chore(categorizeProps): change approach slightly

* chore(AutoExample): extract utility function to dedicated files

* chore(AutoExample): refactor props categorization

also add sorting