chore(ComponentSource): do not set displayName explictly (#576)
* chore(ComponentSource): do not set displayName explictly

this was done to maybe solve something but in turn it broke many code
examples.

* test(StoryPage): remove expectation for code block

this was done in the times when react-docgen-typescript didn't properly
extract displayName. to overcome that, an explicit displayName passed
through story config is used. however, doing that for codeExample breaks
in the case of composed components, making all of composed compnents
have the same name.

explicit displayName can still be used for import example or heading,
but doing it like this for code example is tricky